pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
authors = [
    {name = "Mischan Toosarani-Hausberger", email = "mischan@abovevacant.com"},
]
license = {text = "MIT"}
requires-python = ">=3.14"
dependencies = [
    "typer==0.21.1",
    "sentry-sdk>=2.54.0",
    "google-cloud-storage==3.8.0",
    "protobuf==6.33.5", # TODO: explicitly bump protobuf to the CVE fix until google-cloud-storage bumps
    "pydantic==2.12.5",
    "pandas>=3",
    "deepdiff==8.6.2",
    "textual>=8.2.4",
]
name = "symx"
version = "0.2.0"
description = "Symbol collection for apple platforms."
readme = "README.md"

[project.scripts]
symx = "symx.__main__:main"

[dependency-groups]
dev = [
    "pyright==1.1.408",
    "ruff==0.15.7",
    "pytest==9.0.3",
    "pytest-cov==7.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
